{{cookiecutter.repo_name}}/src/requirements.txt
uwsgi
whitenoise
pillow
coverage
psycopg2
django
django-pipeline
django-grappelli
django-getenv
django-debug-toolbar
